Refactore stor.js
